+++ tests/libs/page-logic.ts
@@ -27,3 +27,21 @@ export async function clickPageMoreActions(page: Page) {
     .getByTestId('editor-option-menu')
     .click();
 }
+
+export async function createPinboardPage(
+  page: Page,
+  parentId: string,
+  title: string
+) {
+  await newPage(page);
+  await page.focus('.affine-default-page-block-title');
+  await page.type('.affine-default-page-block-title', title, {
+    delay: 100,
+  });
+  await clickPageMoreActions(page);
+  await page.getByTestId('move-to-menu-item').click();
+  await page
+    .getByTestId('pinboard-menu')
+    .getByTestId(`pinboard-${parentId}`)
+    .click();
+}
